[refactor] 条件判断

+++ tasks/appium/task_appium_base.py
@@ -227,15 +227,17 @@ class AbsBasicAppiumTask(BaseTask, abc.ABC):
             except Exception:
                 utils_logger.log("release_after_task caught failed")
 
-        # 关闭appium服务
-        utils_common.exec_shell_cmd("ps -ef | grep -v \"$$\" | grep 'appium -p %s -bp %s -U %s' "
-                                    "| awk '{print $2}' | xargs kill -9" %
-                                    (self.appium_port, self.appium_port_bp, self.target_device_name))
-        # 关闭端口占用
-        utils_logger.debug("进程id:", os.getpid(), os.getppid())
-        utils_common.exec_shell_cmd("lsof -i:%s" % self.appium_port)
-        utils_common.exec_shell_cmd("lsof -i:%s | grep -v -E '%s|^COMMAND' | awk '{print $2}' | xargs kill -9" %
-                                    (os.getpid(), self.appium_port))
+        if self.appium_port is not None:
+            if self.target_device_name is not None:
+                # 关闭appium服务
+                utils_common.exec_shell_cmd("ps -ef | grep -v \"$$\" | grep 'appium -p %s -bp %s -U %s' "
+                                            "| awk '{print $2}' | xargs kill -9" %
+                                            (self.appium_port, self.appium_port_bp, self.target_device_name))
+            # 关闭端口占用
+            utils_logger.debug("进程id:", os.getpid(), os.getppid())
+            utils_common.exec_shell_cmd("lsof -i:%s" % self.appium_port)
+            utils_common.exec_shell_cmd("lsof -i:%s | grep -v -E '%s|^COMMAND' | awk '{print $2}' | xargs kill -9" %
+                                        (os.getpid(), self.appium_port))
 
     def is_in_target_page(self, target_page_file, compare_rule=100, retry_count=10, interval_time=0):
